Version 4.5
Bug fixes, mostly, though added one more option:

g:MessageFormatter_formatCurrentLineAsFallback (default 1): if attempting to format a template via the <Plug>FormatCurrentTemplate mapping when not actually
in a template, will fall back to formatting just the current line as an ad-hoc template if this is 1. If 0, will give an error message instead.